Refactoring, linting
