--- src/logger/src/sink/consumer/class.rs
@@ -1,11 +1,0 @@
-use crate::entry::Entry;
-
-
-
-// ================
-// === Consumer ===
-// ================
-
-pub trait Consumer<Levels,Message> {
-    fn consume(&mut self, event:Entry<Levels>, message:Option<Message>);
-}
