+++ src/logger/src/sink/consumer/js_console.rs
@@ -1,7 +1,9 @@
+//! JavaScript console consumer implementation.
+
 use web_sys::console;
 
 use crate::entry::Entry;
-use crate::entry::EntryContent;
+use crate::entry;
 use crate::sink::consumer::Consumer;
 
 
@@ -10,23 +12,24 @@ use crate::sink::consumer::Consumer;
 // === JsConsole Consumer ===
 // ==========================
 
-#[derive(Debug,Default)]
+/// A simple consumer which uses JavaScript console API to print hierarchical logs in a browser.
+#[derive(Clone,Copy,Debug,Default)]
 pub struct JsConsole;
 
 impl<Levels> Consumer<Levels,js_sys::Array> for JsConsole {
     fn consume(&mut self, event:Entry<Levels>, message:Option<js_sys::Array>) {
         match event.content {
-            EntryContent::Message(_) => {
+            entry::Content::Message(_) => {
                 if let Some(msg) = message {
                     console::log(&msg);
                 }
             },
-            EntryContent::GroupBegin(_) => {
+            entry::Content::GroupBegin(_) => {
                 if let Some(msg) = message {
                     console::group_collapsed(&msg);
                 }
             },
-            EntryContent::GroupEnd => {
+            entry::Content::GroupEnd => {
                 console::group_end()
             }
         }
